+++ app/Http/Controllers/AttributeValueController.php
@@ -1,29 +1,52 @@
 <?php
 
 namespace App\Http\Controllers;
-
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use App\Http\Controllers\ImprovedController;
+use App\Models\Attribute;
+use App\Models\AttributeValue;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Log;
 
-class AttributeValueController extends Controller
+class AttributeValueController extends ImprovedController
 {
     public function index($attributeId)
     {
-        $attribute = Attribute::findOrFail($attributeId);
+        $attribute = Attribute::find($attributeId);
+        if (!$attribute) {
+            return $this->respondWithError(['message' => "Attribute not found", 404]);
+        }
+
         $values = $attribute->values;
-        return response()->json($values);
+        if (!$values) {
+            return $this->respondWithError(['message' => "Attribute doesn't have any values"], 404);
+        }
+        return response()->json([$values], 200);
     }
 
     public function show($attributeId, $id)
     {
-        $attribute = Attribute::findOrFail($attributeId);
-        $value = $attribute->values()->findOrFail($id);
-        return response()->json($value);
+        $attribute = Attribute::find($attributeId);
+        if (!$attribute) {
+            return $this->respondWithError(['message' => "Attribute not found"], 404);
+        }
+
+        $value = $attribute->values()->find($id);
+        if (!$value) {
+            return $this->respondWithError(['message' => "Value not found"], 404);
+        }
+
+        return response()->json([$value], 200);
     }
 
     public function store(Request $request, $attributeId)
     {
-        $attribute = Attribute::findOrFail($attributeId);
-        
+        $attribute = Attribute::find($attributeId);
+        if (!$attribute) {
+            return $this->respondWithError(['message' => "Attribute not found"], 404);
+        }
         $validatedData = $request->validate([
             'value' => [
                 'required',
@@ -36,14 +59,24 @@ class AttributeValueController extends Controller
         ]);
 
         $value = $attribute->values()->create($validatedData);
-        return response()->json($value, 201);
+        return response()->json([
+            'message'=> "Value created successfully",
+            'data' => $value
+        ], 201);
     }
 
     public function update(Request $request, $attributeId, $id)
     {
-        $attribute = Attribute::findOrFail($attributeId);
-        $value = $attribute->values()->findOrFail($id);
-        
+        $attribute = Attribute::find($attributeId);
+        if (!$attribute) {
+            return $this->respondWithError(['message' => "Attribute not found"], 404);
+        }
+
+        $value = $attribute->values()->find($id);
+        if (!$value) {
+            return $this->respondWithError(['message' => "Value not found"], 404);
+        }
+
         $validatedData = $request->validate([
             'value' => [
                 'required',
@@ -56,14 +89,26 @@ class AttributeValueController extends Controller
         ]);
 
         $value->update($validatedData);
-        return response()->json($value);
+        return response()->json([
+            'message'=> "Value created successfully",
+            'data' => $value
+        ], 201);
     }
 
     public function destroy($attributeId, $id)
     {
-        $attribute = Attribute::findOrFail($attributeId);
-        $value = $attribute->values()->findOrFail($id);
+        $attribute = Attribute::find($attributeId);
+        if (!$attribute) {
+            return $this->respondWithError(['message' => "Attribute not found"], 404);
+        }
+
+        $value = $attribute->values()->find($id);
+        if (!$value) {
+            return $this->respondWithError(['message' => "Value not found"], 404);
+        }
+
         $value->delete();
-        return response()->json(null, 204);
+        
+        return response()->json(['message' => 'Value deleted successfully'], 200);
     }
 }
